Fix timing tests and fix bug with default `start_time` param
The timing tests that were failing helped discover that new Task objects
were being created with the same `start_time` because the default
parameter `time.time()` is defined/called on initialization of the
class, not the initialization of the object.

The second improvement is writing a `just_wait` method. It continuously
no-ops until the delay elapses. This is better than `time.sleep` because
it does not give up the thread. I was able to improve the time precision
of the unittests by using this new method.